requirements.txt
crewai
crewai-tools
anthropic
python-dotenv
langchain-anthropic
